requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
async-timeout==3.0.1
attrs==19.3.0
black==19.10b0
cfgv==3.1.0
chardet==3.0.4
click==7.1.2
discord.py==1.3.3
distlib==0.3.0
fb-re2==1.0.7
filelock==3.0.12
flake8==3.8.1
identify==1.4.15
idna==2.9
importlib-metadata==1.6.0
mccabe==0.6.1
multidict==4.7.6
nodeenv==1.3.5
pathspec==0.8.0
pkg-resources==0.0.0
pre-commit==2.4.0
pycodestyle==2.6.0
pyflakes==2.2.0
PyYAML==5.3.1
regex==2020.5.14
six==1.14.0
toml==0.10.1
typed-ast==1.4.1
virtualenv==20.0.20
websockets==8.1
yarl==1.4.2
zipp==3.1.0
